M    trunk/makeStr2d.f90 M    trunk/makeStr.f90 *-- fixed a compile error introduced by the last commit.
M    trunk/derivative_structure_generator.f90
*-- slightly updated the concentration check in order to account for
    non-integer number ranges (as calculated from the concentration range).

+++ derivative_structure_generator.f90
@@ -876,7 +876,7 @@ call write_rotperms_list(ParRPList,filename) ! Output might be useful (debugging
 Tcnt = 0 ! Keep track of the total number of structures generated
 HNFcnt = 0 ! Keep track of the total number of symmetrically-inequivalent HNFs in the output
 do ivol = nMin, nMax !max(k,nMin),nMax
-   icRange = ivol*nDFull*cRange
+   icRange = nint(ivol*nDFull*cRange)
    call cpu_time(tstart)
    if (LatDim==3) then !<<< 2D ? or 3D? >>
       call get_all_HNFs(ivol,HNF)  ! 3D
